ai_automation/recommendation_engine/requirements.txt
fastapi
uvicorn
pandas
scipy
scikit-surprise
